// src/app/core/models/inbox.model.ts
//import { FormType } from "../miscellaneous/global-props.template";

export class Inbox {
    assignedById?: string;
    assignedToId?: string;
    dateTime?: Date;
    formId?: string;
   formType?: string;
    id?: number;
    message?: string;
    description?: string;
    readTime?: Date;
    title?: string;
    unread?: boolean;
    assignedByName?: string;
    dateAgoFrame?: string;
}